cli/daemon/run: fix process kill race
If we interrupt `encore run` before the compilation
is finished, we'd sometimes get a panic in `(*Proc).Kill`.

Fix this by checking that the process is running.

+++ cli/daemon/run/proc_groups.go
@@ -411,5 +411,7 @@ func (p *Proc) ProxyReq(w http.ResponseWriter, req *http.Request) {
 // the Process has actually exited. This only kills the Process itself,
 // not any other processes it may have started.
 func (p *Proc) Kill() {
-	_ = p.cmd.Process.Kill()
+	if p.cmd != nil && p.cmd.Process != nil {
+		_ = p.cmd.Process.Kill()
+	}
 }
